systemctl-devicecontroller/requirements.txt
altgraph==0.17.4
ifaddr==0.2.0
importlib-metadata==8.6.1
packaging==24.2
pyserial==3.5
python-dotenv==1.0.1
pyusb==1.3.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
usb==0.0.83.dev0
zipp==3.21.0
